requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
python-dotenv